ssntp: Automatically assign SSNTP client role
We now parse the client certificate and look for
a set of OIDs. We then translate that into an SSNTP
roles bitmask and force the client into using that
role.
The client connection will fail if the requested role
through config does not match the client certificate.

This will allow us to automatically detect which role
a client should use based on its certificate and not expect
clients to explictely specify a role.

Fixes #67

Signed-off-by: Samuel Ortiz <sameo@linux.intel.com>